docs: add comments with relevant references to tix
