+++ src/app/config/config.ts
@@ -90,9 +90,12 @@ const s3 = new aws.S3({
   s3ForcePathStyle: isDev ? true : undefined,
 })
 
-// using aws-sdk v3
+// using aws-sdk v3 (FRM-993)
 const virusScannerLambda = new Lambda({
   region: basicVars.awsConfig.region,
+  // Endpoint is set for development mode to point to the separate docker container running the lambda function.
+  // host.docker.internal is a special DNS name which resolves to the internal IP address used by the host.
+  // Reference: https://docs.docker.com/desktop/networking/#i-want-to-connect-from-a-container-to-a-service-on-the-host
   ...(isDev ? { endpoint: 'http://host.docker.internal:9999' } : undefined),
 })
