+++ src/app/modules/submission/submission.constants.ts
@@ -14,9 +14,8 @@ export const sharedSubmissionParams = {
             .valid(...Object.values(BasicField)),
           answer: Joi.string().allow(''),
           answerArray: Joi.array(),
-          filename: Joi.string(),
+          filename: Joi.string(), // filename or quarantine bucket object key where file has been uploaded to
           content: Joi.binary(),
-          fileKey: Joi.string(), // quarantine bucket object key where file has been uploaded to
           isHeader: Joi.boolean(),
           myInfo: Joi.object(),
           signature: Joi.string().allow(''),
